+++ README.md
@@ -111,14 +111,15 @@ use the github [markdown emoji markup](https://gist.github.com/rxaviers/7360908)
 | :hammer: node-lambda-log            | We want to try a library for [logging](https://github.com/KyleRoss/node-lambda-log), not sure if this is a best library. Will see...                                                                                                          | 03.04.2021 [PR7](https://github.com/mjendza/aws-sandbox/pull/7)   |
 | :hammer: kebab case                 | We want to use kebab case for files and folders convention.                                                                                                                                                                                   | 03.04.2021 [PR7](https://github.com/mjendza/aws-sandbox/pull/7)   |
 | :hammer: pascal case                | For Typescript class names.                                                                                                                                                                                                                   | 03.04.2021 [PR7](https://github.com/mjendza/aws-sandbox/pull/7)   |
-| :hammer: testing                    | We want to use jest library. Based on the friends recommendation this is the best and awesome test library on the javascript market                                                                                                           | 03.04.2021 [PR7](https://github.com/mjendza/aws-sandbox/pull/7)   |
+| :hammer: testing                    | We want to use `jest` library. Based on the friends recommendation this is the best and awesome test library on the javascript market                                                                                                         | 03.04.2021 [PR7](https://github.com/mjendza/aws-sandbox/pull/7)   |
 | :cloud: DynamoDB                    | The context for demo service is users. To minimize latency to 80% cases (read by key) the best fit is use DynamoDB storage. We want to replicate data to all needed regions, so the DynamoDB connection helper calls always local DB instance | 07.04.2021 [PR7](https://github.com/mjendza/aws-sandbox/pull/7)   |
 | :books: cfn-diagram as doc for CDK  | We want to document AWS Stack with diagram based on the CDK                                                                                                                                                                                   | 31.03.2021 [PR9](https://github.com/mjendza/aws-sandbox/pull/9)   |
-| :hammer: database validation        | We decided to use DynamoDB so we don't have any control about stored data (required are key and indexed fields). We want to be sure that data returned by the endpoint are valid, so we put the same ajv validation to the entities.          | 08.04.2021 [PR14](https://github.com/mjendza/aws-sandbox/pull/14) |
+| :hammer: database validation        | We decided to use `DynamoDB` so we don't have any control about stored data (required are key and indexed fields). We want to be sure that data returned by the endpoint are valid, so we put the same ajv validation to the entities.        | 08.04.2021 [PR14](https://github.com/mjendza/aws-sandbox/pull/14) |
 | :hammer: lambda architecture        | We decided to focus on the lambda implementation as proxy integration with API Gateway. All helpers and custom error with error (exception high level handler) are focused to deliver data as Lambda Proxy Integration.                       | 13.04.2021 [PR14](https://github.com/mjendza/aws-sandbox/pull/14) |
 | :cloud: CDK resource names and id's | CDK resources should have id generated by `generateResourceName` helper and autogenerated name. Use `CdkResources` interface to define all needed AWS objects.                                                                                | 14.04.2021 [PR14](https://github.com/mjendza/aws-sandbox/pull/14) |
 | :hammer: delivering features        | Push to develop branch is forbidden. Use PR with squash commits to keep clean main branch                                                                                                                                                     | 16.04.2021 [PR1](https://github.com/mjendza/aws-sandbox/pull/1)   |
 | :cloud: Event-Driven Architecture   | I really don't like coupling! To build lousily coupled architecture in the cloud we can use EventBridge. Check dedicated [Event-Driven-Architecture paragraf](#Event-Driven-Architecture).                                                    | 04.05.2021 [PR22](https://github.com/mjendza/aws-sandbox/pull/22) |
+| :hammer: lambda architecture        | We decided to remove N-Layer architecure - code should be simple.                                                                                                                                                                             | 25.06.2022 [PR14](https://github.com/mjendza/aws-sandbox/pull/XX) |
 
 ## architecture
 
@@ -138,10 +139,11 @@ To deliver lousily coupled architecture the best option is to use the EventBridg
 
 ### lambda
 
-The lambda is written in N-Layer. This is a microservice, so logic should be so 'easy' and hope in this can be migrated without any refactoring into hexagonal architecture.  
+### 
+The lambda is written as simple as possible. This is a microservice, so logic should be so 'easy' and hope in this can be migrated without any refactoring into hexagonal architecture.  
 On the picture there is a flow inside the lambda. There is a Service Layer with the dedicated model from the Lambda event. For the persistence layer is a repository to communicate (makes) requests to the DynamoDB.
 
-![n-layer](doc/solution/lambda-architecture-Lambda_Architecture_N_Layer.png)
+![KISS-lambda](doc/solution/lambda-architecture.png)
 
 #### helpers
