+++ bin/src/handlers/create-user/index.ts
@@ -7,7 +7,6 @@ import { EventBridgeEvent } from 'aws-lambda';
 import * as log from 'lambda-log';
 import { dynamoClient } from '../../helpers/dynamodb-factory';
 import { CreateUserService } from '../../users/create-user-service';
-import { UserRepository } from '../../users/user-repository';
 
 export const handler = async (
     event: EventBridgeEvent<string, any>
@@ -19,9 +18,7 @@ export const handler = async (
             createUserEventSchema
         );
         log.info(`model: ${JSON.stringify(model)}`);
-        const service = new CreateUserService(
-            new UserRepository(dynamoClient())
-        );
+        const service = new CreateUserService(dynamoClient());
         const result = await service.create(model.detail);
         log.info(`result: ${JSON.stringify(result)}`);
     } catch (error) {
