--- bin/src/users/user-repository.ts
@@ -1,49 +1,0 @@
-import { UserEntity, userEntitySchema } from './user-entity';
-import { validateEntity } from '../helpers/validation-helpers';
-import { CreateUserHandlerLambdaSettings } from '../../../cdk/core/settings/lambda-settings';
-import { DynamoDB } from 'aws-sdk';
-import * as log from 'lambda-log';
-import { DocumentClient } from 'aws-sdk/clients/dynamodb';
-import { LambdaProxyError } from '../helpers/lambda-proxy-error';
-import { getEnvironmentSettingsKey } from '../helpers/get-environment-settings-key';
-
-export class UserRepository {
-    private tableName =
-        getEnvironmentSettingsKey<CreateUserHandlerLambdaSettings>(
-            'USER_TABLE_NAME'
-        );
-    static IndexSeparator: string = '_';
-    constructor(private documentClient: DynamoDB.DocumentClient) {}
-
-    async put(item: UserEntity): Promise<void> {
-        const params = {
-            TableName: this.tableName,
-            Item: item,
-        };
-        log.info(`DynamoDB params: ${JSON.stringify(params)}`);
-        try {
-            await this.documentClient.put(params).promise();
-        } catch (dbError) {
-            log.error(`DynamoDB ERROR: ${JSON.stringify(dbError)}`);
-            throw new Error(`DynamoDB ERROR: ${dbError}`);
-        }
-    }
-
-    async get(id: string): Promise<UserEntity> {
-        const params: DocumentClient.GetItemInput = {
-            TableName: this.tableName,
-            Key: {
-                ['id']: id,
-            },
-        };
-        log.info(`DynamoDB params: ${JSON.stringify(params)}`);
-
-        const result = await this.documentClient.get(params).promise();
-        if (!result.Item) {
-            throw new LambdaProxyError(404, 'User does not exists.');
-        }
-        log.info(`DynamoDB result: ${JSON.stringify(result.Item)}`);
-        const model = validateEntity<UserEntity>(result.Item, userEntitySchema);
-        return model;
-    }
-}
