+++ doc/solution/event-driven-architecutre.puml
@@ -1,5 +1,5 @@
 @startuml Event Driven Architecture
-!includeurl https://raw.githubusercontent.com/RicardoNiepel/C4-PlantUML/release/1-0/C4_Container.puml
+!includeurl https://raw.githubusercontent.com/RicardoNiepel/C4-PlantUML/master/C4_Container.puml
 
 title
 Event Driven Architecture
